Add lots of dependency and docblock fixes to work with Psalm
Psalm is a static type checker for PHP that's under development.
It will be open sourced soonish, at which point it will be added
to the tests for this repo. This commit contains lots of fixes
found by Psalm, as well afs some notations to help it do its job.