Allow a customer to be created as part of a purchase request
I thought you had to create the customer first, but I was wrong.
If you provide a customer ID that doesn't exist, Vindicia will
create it for you. This works for authorize, purchase, and create
subscription. Added functionality to specify the customer name
and email at this time as well. Updated the documentation
accordingly.