Allow a saved PayPal payment method to be set on a subscription (#27)
Currently if you set a saved payment method ID with the
CreateSubscriptionRequest, it always sets the type to credit card,
so then Vindicia tries to validate the credit card and fails. Now,
if you just provide a payment method ID and no other payment method
details, a type won't be set so it will pass.